DRA func_80130264 and func_80130618
