+++ config/symbols.hd.dra.txt
@@ -83,6 +83,16 @@ D_800B0914 = 0x800B09D8;
 D_800B0918 = 0x800B09DC;
 D_800B091C = 0x800B09E0;
 D_800B0920 = 0x800B09E4;
+D_800B0A4C_hd = 0x800B0A4C;
+D_800B0A5C_hd = 0x800B0A5C;
+D_800B0A6C_hd = 0x800B0A6C;
+D_800B0A7C_hd = 0x800B0A7C;
+D_800B0A8C_hd = 0x800B0A8C;
+D_800B0A9C_hd = 0x800B0A9C;
+D_800B0AAC_hd = 0x800B0AAC;
+D_800B0A3C = 0x800B0B00;
+D_800B0A4C = 0x800B0B10;
+D_800B0AAC = 0x800B0B70;
 D_800B0CB4 = 0x800B0D78;
 D_800B0F94 = 0x800B1058;
 aPbav = 0x800B1140;
@@ -205,7 +215,6 @@ TimeAttackController = 0x800FD474;
 func_800FD6C4 = 0x800FD678;
 AddToInventory = 0x800FD828;
 GetStatusAilmentTimer = 0x800FDACC;
-CastSpell = 0x800FDC48;
 IsRelicActive = 0x800FE360;
 func_800FF0F4 = 0x800FF0A0;
 HandleSaveMenu = 0x801030C8;
